refactored code in IFrame component
