+++ src/components/main/stageView/iFrame/helpers/handleAnchorTagDoubleClick.ts
@@ -1,0 +1,7 @@
+export const handleAnchorTagDoubleClick = (
+	ele: any
+	) => {
+	if (ele.tagName === "A" && ele.href ) {
+	  // window.open(ele.href, '_blank', 'noreferrer'); // Issue: 238
+	}
+  };
