+++ src/components/main/stageView/iFrame/helpers/index.ts
@@ -1,0 +1,5 @@
+export * from './createAndInsertElement';
+export * from './cloneAndInsertNode';
+export * from './handleAnchorTagDoubleClick';
+export * from './handleEditableElementDoubleClick';
+export * from './handleWebComponentDblClick';
